tests/wpt/tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-shorthand-serialization.html

<!DOCTYPE html>
<html>
<head>
  <title>CSS Grid Lanes: grid-lanes serializes properly when longhands are set</title>
  <link rel="author" title="Yanling Wang" href="mailto:yanlingwang@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
    function testValidGridLanes(gridTemplateRowsValue, gridTemplateColumnsValue, gridTemplateAreasValue, gridLanesDirectionValue, serializedGridLanesValue) {
    test(()=>{
      const root = document.documentElement;
    const properties = [
      ["gridTemplateRows", gridTemplateRowsValue],
      ["gridTemplateColumns", gridTemplateColumnsValue],
      ["gridTemplateAreas", gridTemplateAreasValue],
      ["gridLanesDirection", gridLanesDirectionValue],
    ];
    for (const [property, value] of properties) {
      root.style[property] = "";
      root.style[property] = value;
    }
    assert_equals(root.style.gridLanes, serializedGridLanesValue);
  }, `grid-template-rows: ${gridTemplateRowsValue},
      grid-template-columns: ${gridTemplateColumnsValue},
      grid-template-areas: ${gridTemplateAreasValue},
      grid-lanes-direction: ${gridLanesDirectionValue} should be valid.`);
}

  testValidGridLanes("none", "none", "none", "column", "none column");
  testValidGridLanes("10px", "none", "none", "column", "none column");
  testValidGridLanes("10px 20px", "10% 20%", "none", "row", "10px 20px row");
  testValidGridLanes("none", "1fr 1fr 3fr", '"a a b"', 'column', '"a a b" 1fr 1fr 3fr column');
  testValidGridLanes("20% 40%", "none", '"b" "a"', 'row', '"b a" 20% 40% row');
  testValidGridLanes("none", "fit-content(calc(0.5em + 10px))", "none", "column track-reverse fill-reverse", "fit-content(calc(0.5em + 10px)) column track-reverse fill-reverse");
  testValidGridLanes("10% 20% 40%", "none", '"a" "b" "c"', "row fill-reverse track-reverse", '"a b c" 10% 20% 40% row fill-reverse track-reverse');
  </script>
</body>
</html>
